Call Checks

A set of helper functions that determine what type of call is being made within an expression.

is_reactive_call checks whether or not the call is evaluating a reactive variable.

is_reactive_val_call checks whether or not the call is evaluating a reactiveVal variable.

is_reactive_values_call checks whether or not the call is evaluating an item within a reactiveValues variable.

is_any_reactive_call checks whether or not the call points to evaluating a reactive, reactiveVal or reactiveValues.

is_variable_call checks whether or not the call point to a variable that is defined within the given module.

is_input_call checks whether or not the call points to evaluate an input value.

is_session_user_data checks whether or not the call points to evaluate an object within session$userData

Usage

is_reactive_call(x, env = rlang::caller_env())

is_reactive_val_call(x, env = rlang::caller_env())

is_reactive_values_call(x, env = rlang::caller_env())

is_any_reactive_call(x, env = rlang::caller_env())

is_variable_call(x, existing_vars = NULL, env = rlang::caller_env())

is_input_call(x)

is_session_user_data(x)

Arguments

x

An R call object

env

The environment the call is being made, by default it is the environment calling the check, but is likely the environment the call is being made i.e. the reactive expression.

existing_vars

A character vector of variable definitions that exist in the Repro object

Value

A boolean value determining whether or not the call check has passed.